🔧 refactor nft storage in services
