+++ components/bsx/Create/CreateCollection.vue
@@ -49,7 +49,7 @@ import MetaTransactionMixin from '@/utils/mixins/metaMixin'
 import PrefixMixin from '@/utils/mixins/prefixMixin'
 import ApiUrlMixin from '@/utils/mixins/apiUrlMixin'
 import { notificationTypes, showNotification } from '@/utils/notification'
-import { pinJson, PinningKey } from '@/utils/nftStorage'
+import { pinJson, PinningKey } from '~/services/nftStorage'
 import resolveQueryPath from '@/utils/queryPathResolver'
 import { getImageTypeSafe, pinImageSafe } from '@/utils/safePin'
 import { estimate } from '@/utils/transactionExecutor'
