+++ components/unique/Create/CreateCollection.vue
@@ -34,7 +34,7 @@ import { unSanitizeIpfsUrl } from '@kodadot1/minimark'
 import AuthMixin from '@/utils/mixins/authMixin'
 import MetaTransactionMixin from '@/utils/mixins/metaMixin'
 import { notificationTypes, showNotification } from '@/utils/notification'
-import { pinFileToIPFS, pinJson, PinningKey } from '@/utils/nftStorage'
+import { pinFileToIPFS, pinJson, PinningKey } from '~/services/nftStorage'
 import { canSupport } from '@/utils/support'
 import { estimate, Extrinsic } from '@/utils/transactionExecutor'
 import { createMetadata } from '@kodadot1/minimark'
